Added base class for date detection inside date
